requirements.txt
Django==1.6.6
django-disposable-email-checker==0.2.3
httplib2==0.9
sorl-thumbnail==11.12
wsgiref==0.1.2
South==0.8.4
dj-database-url==0.3.0
dj-static==0.0.5
django-cors-headers==0.12
django-toolbelt==0.0.1
django-filter==0.7
gunicorn==19.0.0
psycopg2==2.5.3
pystache==0.5.4
static==1.0.2
